requirements.txt
torch>=2.5.1
fair-esm>=2.0.0
biopython>=1.86
networkx>=3.6.1
numpy>=2.4.4
pandas>=3.0.1
scipy>=1.17.1
tqdm>=4.67.3